doc: inform users that fuses are blown *before* shipping
The documentation is extremely vague about exactly *who* blows the
fuses on the Tillitis.  Based on your pre-release hardware and
common sense about user-controlled hardware, I and several others
assumed that the customer is the one who does this.  It was quite a
rude surprise to discover that this is not the case.

You should be more hoenst about this, particularly at the time that
orders are placed from your shop.