fw: Remove TK1_MMIO_TK1_BLAKE2S and its use
Remove mentions of TK1_MMIO_TK1_BLAKE2S and its use in firmware, where
we used to expose the address of the blake2s() function in firmware.

Since the introduction of the syscall mechanism we don't allow
execution in ROM anymore so it's impossible to call the firmware's
blake2s() function.