fw: Change to two modes of CDI computation
Change to using two modes for the CDI computation:

1. As before, based on the UDS, the hash of the app, and optionally
   the USS. This is the default.

2. If the seed_digest in struct reset is set (indicated by setting the
   new struct reset member mask bitmask to include RESET_SEED) compute
   CDI like this instead:

   CDI = blake2s(UDS, seed_digest[, USS])

   This way we give the next app a deterministic secret seed for key
   material without depending on the app binary not changing.

The seed_digest used in number 2 should be filled in by the running
app when doing the sys_reset() system call. Firmware will mix this
with the running app's CDI like

  seed_digest = blake2s(CDI, seed_digest)

before actually doing the reset. After the reset, the CDI is computed
as above, using the seed_digest.

This means that if we chain apps after each other and want to have a
combination of measured boot and verified boot, we can run a boot
verifier first, let it verify the next app, then create a seed for a
secret and deliver it to the next app through this method. This
solves:

- the next app gets a deterministic seed to base it's key material on,
  which means it can be updated without losing key material.

- the seed is delivered through the usual CDI, so no updates are
  needed to the app to handle both the case where it's the first app to
  start or if it's part of a chain.

- the seed is always tied to some data the caller wanted to mix
  in (like a vendor pubkey and a name, perhaps), and the CDI of the app
  that wants the next app to boot, establishing a chain in the creation
  of key material.

- the new CDI is unknown to the caller. The old CDI is unknown to the
  receiver.