ch552: add zero length packet handling
Terminate all 64-byte frames with a zero-length packet. Handle incoming
zero-byte frames by ignoring them. This fix addresses issues related to
transfers of an exact length of 64 bytes (or multiples thereof) on all
platforms. For simplicity, all 64-byte frames are terminated, regardless
of whether they are the last or not.