fw: Calculate different CDI depending on how an app was loaded
A `domain` byte is added to the data that is hashed to produce the CDI.
The value of the domain byte mirrors the value that is present in
`resetinfo->type`.

This makes it harder for an app loaded from one source to impersonate an
app loaded from another source. An app loaded from the client will not
get the same CDI as an app loaded from flash.